Hien Pust Int

+++ FoodOrderingSystem_Client/FoodOrderingSystem/app/src/main/java/com/example/foodorderingsystem/Activity/FoodDetail.java
@@ -27,7 +27,7 @@ public class FoodDetail extends AppCompatActivity {
         Intent intent = getIntent();
 
         pName = intent.getStringExtra("name");
-        pPrice = intent.getDoubleExtra("price", 0);
+        pPrice = intent.getIntExtra("price", 0);
         pDescription = intent.getStringExtra("description");
         iURL = intent.getStringExtra("image");
